pyproject.toml
[project]
name = "bi-agent-retailware"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu==1.11.0",
    "fastapi==0.115.9",
    "langchain==0.3.25",
    "langchain-community==0.3.21",
    "langchain-openai==0.2.10",
    "langgraph==0.2.50",
    "pandas==2.2.3",
    "pydantic==2.10.4",
    "pyodbc==5.2.0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
    "sqlalchemy==2.0.41",
    "uvicorn==0.32.1",
]
